+++ source/release-notes.md
@@ -848,3 +848,6 @@
 05.03.2019
 * health-check: add failed login attempts information on daily HTML report
 * maintenance-plan: add option to trigger a log backup and log shrink (truncate only) after index maintenance
+
+11.03.2019
+* maintenance-plan: change indexes for [maintenance-plan].[objectSkipList] table